refactor: applies review comments.
