+++ services/docker/docker-cloud-automated.service.js
@@ -1,34 +1,19 @@
 'use strict'
 
-const Joi = require('joi')
 const { BaseJsonService } = require('..')
 const { dockerBlue, buildDockerUrl } = require('./docker-helpers')
-
-const cloudAutomatedBuildSchema = Joi.object({
-  objects: Joi.array()
-    .items(Joi.object({ build_settings: Joi.array() }).required())
-    .required(),
-}).required()
+const { fetchBuild } = require('./docker-cloud-common-fetch')
 
 module.exports = class DockerCloudAutomatedBuild extends BaseJsonService {
-  async fetch({ user, repo }) {
-    return this._requestJson({
-      schema: cloudAutomatedBuildSchema,
-      url: `https://cloud.docker.com/api/build/v1/source/?image=${user}/${repo}`,
-      errorMessages: { 404: 'repo not found' },
-    })
-  }
-
   static render({ buildSettings }) {
     if (buildSettings.length >= 1) {
       return { message: 'automated', color: dockerBlue }
-    } else {
-      return { message: 'manual', color: 'yellow' }
     }
+    return { message: 'manual', color: 'yellow' }
   }
 
   async handle({ user, repo }) {
-    const data = await this.fetch({ user, repo })
+    const data = await fetchBuild(this, { user, repo })
     return this.constructor.render({
       buildSettings: data.objects[0].build_settings,
     })
@@ -49,7 +34,9 @@ module.exports = class DockerCloudAutomatedBuild extends BaseJsonService {
   static get examples() {
     return [
       {
-        title: 'Docker Automated build (new hub)',
+        title: 'Docker Cloud Automated build',
+        documentation:
+          '<p>For the new Docker Hub (https://cloud.docker.com)</p>',
         namedParams: {
           user: 'jrottenberg',
           repo: 'ffmpeg',
